+++ controllers/dogu_registrator_test.go
@@ -36,6 +36,7 @@ func TestEtcdDoguRegistrator_RegisterDogu(t *testing.T) {
 	testErr := errors.New("test")
 
 	t.Run("successfully register a dogu", func(t *testing.T) {
+		// given
 		registryMock := &cesmocks.Registry{}
 		doguRegistryMock := &cesmocks.DoguRegistry{}
 		doguConfigMock := &cesmocks.ConfigurationContext{}
@@ -46,51 +47,57 @@ func TestEtcdDoguRegistrator_RegisterDogu(t *testing.T) {
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		registryMock.Mock.On("DoguConfig", mock.Anything).Return(doguConfigMock)
 		doguResourceGenerator.Mock.On("GetDoguSecret", mock.Anything, mock.Anything).Return(&v1.Secret{ObjectMeta: metav1.ObjectMeta{Name: "ldap-private", Namespace: "clusterns"}}, nil)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, doguResourceGenerator)
 
+		// when
 		err := registrator.RegisterDogu(ctx, ldapCr, ldapDogu)
-		require.NoError(t, err)
 
+		// then
+		require.NoError(t, err)
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock, doguConfigMock)
 	})
 
 	t.Run("fail to register dogu", func(t *testing.T) {
+		// given
 		registryMock := &cesmocks.Registry{}
 		doguRegistryMock := &cesmocks.DoguRegistry{}
 		doguResourceGenerator := &mocks.DoguResourceGenerator{}
 		doguRegistryMock.Mock.On("Register", mock.Anything).Return(testErr)
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, doguResourceGenerator)
 
+		// when
 		err := registrator.RegisterDogu(ctx, ldapCr, ldapDogu)
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to register dogu")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock)
 
 	})
 
 	t.Run("fail to enable dogu", func(t *testing.T) {
+		// given
 		registryMock := &cesmocks.Registry{}
 		doguRegistryMock := &cesmocks.DoguRegistry{}
 		doguResourceGenerator := &mocks.DoguResourceGenerator{}
 		doguRegistryMock.Mock.On("Register", mock.Anything).Return(nil)
 		doguRegistryMock.Mock.On("Enable", mock.Anything).Return(testErr)
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, doguResourceGenerator)
 
+		// when
 		err := registrator.RegisterDogu(ctx, ldapCr, ldapDogu)
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to enable dogu")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock)
 
 	})
 
 	t.Run("fail to write public key", func(t *testing.T) {
+		// given
 		registryMock := &cesmocks.Registry{}
 		doguRegistryMock := &cesmocks.DoguRegistry{}
 		doguConfigMock := &cesmocks.ConfigurationContext{}
@@ -100,16 +107,18 @@ func TestEtcdDoguRegistrator_RegisterDogu(t *testing.T) {
 		doguRegistryMock.Mock.On("Enable", mock.Anything).Return(nil)
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		registryMock.Mock.On("DoguConfig", mock.Anything).Return(doguConfigMock)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, doguResourceGenerator)
 
+		// when
 		err := registrator.RegisterDogu(ctx, ldapCr, ldapDogu)
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to write")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock, doguConfigMock)
 	})
 	t.Run("fail generate secret", func(t *testing.T) {
+		// given
 		scheme := runtime.NewScheme()
 		client := fake.NewClientBuilder().WithScheme(scheme).Build()
 		registryMock := &cesmocks.Registry{}
@@ -122,17 +131,19 @@ func TestEtcdDoguRegistrator_RegisterDogu(t *testing.T) {
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		registryMock.Mock.On("DoguConfig", mock.Anything).Return(doguConfigMock)
 		doguResourceGenerator.Mock.On("GetDoguSecret", mock.Anything, mock.Anything).Return(nil, testErr)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, doguResourceGenerator)
 
+		// when
 		err := registrator.RegisterDogu(ctx, ldapCr, ldapDogu)
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to generate secret")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock, doguConfigMock)
 	})
 
 	t.Run("fail create secret", func(t *testing.T) {
+		// given
 		scheme := runtime.NewScheme()
 		client := fake.NewClientBuilder().WithScheme(scheme).Build()
 		registryMock := &cesmocks.Registry{}
@@ -145,12 +156,13 @@ func TestEtcdDoguRegistrator_RegisterDogu(t *testing.T) {
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		registryMock.Mock.On("DoguConfig", mock.Anything).Return(doguConfigMock)
 		doguResourceGenerator.Mock.On("GetDoguSecret", mock.Anything, mock.Anything).Return(&v1.Secret{ObjectMeta: metav1.ObjectMeta{Name: "ldap-private", Namespace: "clusterns"}}, nil)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, doguResourceGenerator)
 
+		// when
 		err := registrator.RegisterDogu(ctx, ldapCr, ldapDogu)
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to create secret")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock, doguConfigMock)
 	})
@@ -160,6 +172,7 @@ func TestCESDoguRegistrator_UnregisterDogu(t *testing.T) {
 	testErr := errors.New("test")
 
 	t.Run("successfully unregister a dogu", func(t *testing.T) {
+		// given
 		scheme := runtime.NewScheme()
 		client := fake.NewClientBuilder().WithScheme(scheme).Build()
 		registryMock := &cesmocks.Registry{}
@@ -169,16 +182,18 @@ func TestCESDoguRegistrator_UnregisterDogu(t *testing.T) {
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		doguConfigMock.Mock.On("RemoveAll").Return(nil)
 		doguRegistryMock.Mock.On("Unregister", mock.Anything).Return(nil)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, &ResourceGenerator{})
 
+		// when
 		err := registrator.UnregisterDogu("ldap")
-		require.NoError(t, err)
 
+		// then
+		require.NoError(t, err)
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock)
 	})
 
 	t.Run("failed to remove dogu config", func(t *testing.T) {
+		// given
 		scheme := runtime.NewScheme()
 		client := fake.NewClientBuilder().WithScheme(scheme).Build()
 		registryMock := &cesmocks.Registry{}
@@ -187,17 +202,19 @@ func TestCESDoguRegistrator_UnregisterDogu(t *testing.T) {
 		registryMock.Mock.On("DoguConfig", mock.Anything).Return(doguConfigMock)
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		doguConfigMock.Mock.On("RemoveAll").Return(testErr)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, &ResourceGenerator{})
 
+		// when
 		err := registrator.UnregisterDogu("ldap")
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to remove dogu config")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock, doguConfigMock)
 	})
 
 	t.Run("failed to unregister dogu", func(t *testing.T) {
+		// given
 		scheme := runtime.NewScheme()
 		client := fake.NewClientBuilder().WithScheme(scheme).Build()
 		registryMock := &cesmocks.Registry{}
@@ -207,12 +224,13 @@ func TestCESDoguRegistrator_UnregisterDogu(t *testing.T) {
 		registryMock.Mock.On("DoguRegistry").Return(doguRegistryMock)
 		doguConfigMock.Mock.On("RemoveAll").Return(nil)
 		doguRegistryMock.Mock.On("Unregister", mock.Anything).Return(testErr)
-
 		registrator := NewCESDoguRegistrator(client, registryMock, &ResourceGenerator{})
 
+		// when
 		err := registrator.UnregisterDogu("ldap")
-		require.Error(t, err)
 
+		// then
+		require.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to unregister dogu")
 		mock.AssertExpectationsForObjects(t, registryMock, doguRegistryMock, doguConfigMock)
 	})